feat: pnpm db:validate compares a local backup against the live database
Read-only doc-by-doc comparison (stable-stringified, order-insensitive)
reporting docs only in the backup, only in the database, or with different
contents; exit 1 on any mismatch. Shared dump/load logic moved into
scripts/lib/firestoreBackupShared.js for backup/restore/validate.
