+++ package.json
@@ -51,7 +51,8 @@
     "format:check": "prettier --check \"src/**/*.{js,jsx,ts,tsx,css}\"",
     "emulators": "firebase emulators:start --only auth,firestore $([ -d .emulator-data ] && echo '--import=.emulator-data') --export-on-exit=.emulator-data",
     "db:backup": "node scripts/firestore-backup.js",
-    "db:restore": "node scripts/firestore-restore.js"
+    "db:restore": "node scripts/firestore-restore.js",
+    "db:validate": "node scripts/firestore-validate.js"
   },
   "eslintConfig": {
     "extends": [
